+++ src/scheme/ws.cpp
@@ -51,7 +51,7 @@ void awh::server::scheme::Websocket::set(const uint32_t bid) noexcept {
 			// Устанавливаем список доступных компрессоров
 			ret.first->second->http.req.compressors(this->compressors);
 			// Устанавливаем контрольную точку
-			ret.first->second->respPong = this->_fmk->timestamp <uint32_t> (fmk_t::chrono_t::MILLISECONDS);
+			ret.first->second->respPong = this->_fmk->timestamp <uint64_t> (fmk_t::chrono_t::MILLISECONDS);
 		}
 	}
 }
